requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
django-iranian-cities==2.0.0
django-jalali==7.3.0
djangorestframework==3.15.2
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
kavenegar==1.1.2
pillow==11.1.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
